requirements.txt
numpy
pandas
scikit-learn
fastapi
uvicorn
pytest
requests
pyyaml